chore: remove aws-lambda dependency
The `aws-lambda` package is actually a CLI, unrelated to imports of the
`"aws-lambda"` path.

At run-time, AWS will make the `aws-lambda` module available (meaning it
doesn't need to be installed as an NPM dependency, and can't be).

The `@types/aws-lambda` package does however provide types for the
actual aws-lambda module.